raarchart added as a component, tooltip and fill added to svgs on radar chart
